backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.9.2
starlette==0.38.5
python-multipart==0.0.9



